表重新生成

+++ src/main/java/com/xiangxing/controller/ExamController.java
@@ -50,7 +50,7 @@ public class ExamController extends BaseController {
 	}
 
 	@RequestMapping("/destroyExam")
-	public void destroyexam(Integer examId) {
+	public void destroyexam(Long examId) {
 		examMapper.deleteByPrimaryKey(examId);
 		writeToOkResponse();
 	}
